Address review findings and fix agent dialog race
